gitignore

+++ server/controller/user.js
@@ -4,10 +4,10 @@ exports.getUsers = [
   async (req, res) => {
     try {
       const response = await User.findAll();
-      res.status(200).json({ success: true, message: response });
+      return res.status(200).json({ success: true, message: response });
     } catch (error) {
       console.log(error);
-      res.status(400).json({ success: false, message: error.message });
+      return res.status(400).json({ success: false, message: error.message });
     }
   },
 ];
@@ -16,10 +16,15 @@ exports.getUserById = [
   async (req, res) => {
     try {
       const response = await User.findOne({ where: { id: req.params.id } });
-      res.status(200).json({ success: true, message: response });
+      if (!response)
+        return res
+          .status(400)
+          .json({ success: false, message: "User Doesnt Exist" });
+
+      return res.status(200).json({ success: true, message: response });
     } catch (error) {
       console.log(error);
-      res.status(400).json({ success: false, message: error.message });
+      return res.status(400).json({ success: false, message: error.message });
     }
   },
 ];
@@ -27,11 +32,25 @@ exports.getUserById = [
 exports.createUsers = [
   async (req, res) => {
     try {
+      const usernameExist = await User.findOne({
+        where: { name: req.body.name },
+      });
+      const emailExist = await User.findOne({
+        where: { email: req.body.email },
+      });
+      if (usernameExist || emailExist)
+        return res
+          .status(400)
+          .json({ success: false, message: "Username or Email already Exist" });
+      if (!req.body.name || !req.body.email || !req.body.gender)
+        return res
+          .status(400)
+          .json({ success: false, message: "Please Fill Out All The Field" });
       await User.create(req.body);
-      res.status(201).json({ success: true, message: "User Created" });
+      return res.status(201).json({ success: true, message: "User Created" });
     } catch (error) {
       console.log(error);
-      res.status(400).json({ success: false, message: error.message });
+      return res.status(400).json({ success: false, message: error.message });
     }
   },
 ];
@@ -39,15 +58,25 @@ exports.createUsers = [
 exports.updateUsers = [
   async (req, res) => {
     try {
+      const userExist = await User.findOne({ where: { id: req.params.id } });
+      if (!userExist)
+        return res
+          .status(400)
+          .json({ success: false, message: "User Doesn`t Exist" });
+      if (!req.body)
+        return res.status(400).json({
+          success: false,
+          message: "Please Fill Out The Field You Want To Update",
+        });
       await User.update(req.body, {
         where: {
           id: req.params.id,
         },
       });
-      res.status(200).json({ success: true, message: "User Updated" });
+      return res.status(200).json({ success: true, message: "User Updated" });
     } catch (error) {
       console.log(error);
-      res.status(400).json({ success: false, message: error.message });
+      return res.status(400).json({ success: false, message: error.message });
     }
   },
 ];
@@ -55,15 +84,21 @@ exports.updateUsers = [
 exports.deleteUsers = [
   async (req, res) => {
     try {
+      const userExist = await User.findOne({ where: { id: req.params.id } });
+      if (!userExist)
+        return res
+          .status(400)
+          .json({ success: false, message: "User Doesnt Exist" });
+
       await User.destroy({
         where: {
           id: req.params.id,
         },
       });
-      res.status(200).json({ success: true, message: "User Deleted" });
+      return res.status(200).json({ success: true, message: "User Deleted" });
     } catch (error) {
       console.log(error);
-      res.status(400).json({ success: false, message: error.message });
+      return res.status(400).json({ success: false, message: error.message });
     }
   },
 ];
